Make a first step of the big refactoring
The previous program structure was cumbersome to work with so this is
another attempt. As part of that a lot of funcionality was lost. That's
because it have to start simple. New features can be added once they are
well thought out.

The approach introduced by this CL is more production ready and is
inspired by the Acme editor source code. Current implementation
supports line wrapping of long lines and temporarily loses any vi
related commands.